rename enum to avoid python confusion
