+++ museek/plugin/demo/demo_flip_plugin.py
@@ -4,7 +4,7 @@ from PIL.Image import Transpose
 from ivory.plugin.abstract_plugin import AbstractPlugin
 from ivory.utils.requirement import Requirement
 from ivory.utils.result import Result
-from museek.enum.demo_enum import DemoEnum
+from museek.enums.demo_enum import DemoEnum
 
 
 class DemoFlipPlugin(AbstractPlugin):
