+++ museek/plugin/out_plugin.py
@@ -4,7 +4,7 @@ from definitions import ROOT_DIR
 from ivory.plugin.abstract_plugin import AbstractPlugin
 from ivory.utils.requirement import Requirement
 from ivory.utils.result import Result
-from museek.enum.result_enum import ResultEnum
+from museek.enums.result_enum import ResultEnum
 
 
 class OutPlugin(AbstractPlugin):
