+++ museek/plugin/point_source_flagger_plugin.py
@@ -1,7 +1,7 @@
 from ivory.plugin.abstract_plugin import AbstractPlugin
 from ivory.utils.requirement import Requirement
 from ivory.utils.result import Result
-from museek.enum.result_enum import ResultEnum
+from museek.enums.result_enum import ResultEnum
 from museek.flag_factory import FlagFactory
 from museek.time_ordered_data import TimeOrderedData
