+++ museek/plugin/scan_track_split_plugin.py
@@ -5,8 +5,8 @@ from definitions import ROOT_DIR
 from ivory.plugin.abstract_plugin import AbstractPlugin
 from ivory.utils.requirement import Requirement
 from ivory.utils.result import Result
-from museek.enum.result_enum import ResultEnum
-from museek.enum.scan_state_enum import ScanStateEnum
+from museek.enums.result_enum import ResultEnum
+from museek.enums.scan_state_enum import ScanStateEnum
 from museek.time_ordered_data import TimeOrderedData
